# Conflicts:
#	tests/CMakeLists.txt
#	tests/Makefile.am

+++ example/tcpip-forward.c
@@ -229,7 +229,7 @@ int main(int argc, char *argv[])
     }
 
     sin.sin_family = AF_INET;
-    sin.sin_port = htons(local_destport);
+    sin.sin_port = htons((unsigned short)local_destport);
     sin.sin_addr.s_addr = inet_addr(local_destip);
     if(INADDR_NONE == sin.sin_addr.s_addr) {
         perror("inet_addr");
